requirements.txt
numpy>=1.19.0
torch>=1.9.0
transformers>=4.5.0
sentence-transformers>=2.2.0
scikit-learn>=0.24.0
flask>=2.0.0
pymongo>=4.0.0
networkx>=2.6.0
faiss-cpu>=1.7.0
spacy>=3.0.0
python-dotenv>=0.19.0
pandas>=1.3.0
scipy>=1.7.0
tensorflow>=2.8.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
gunicorn>=20.1.0